Fix problems downloading larger files
The previous mechanism for downloading the file did not cope well on Chrome when downloading logs with higher (but not particularly large) numbers of lines, e.g. 25,000 lines. Trying to download a file in Chrome with 25,000 lines would result in "Failed - Network Error" in Chrome.

This change alters the download function to create a Blob and create an object URL from the blob. I have not tested this in IE, but the API calls should working on IE10 or above.